Add functionality for customizing an invite
Please note due to several GitPod issues and grabbing time off-line
when I could while home schooling, most of the JS/CSS for this
functionality was written off-line and just tested locally on my
laptop. Hence this initial large commit of the JS, although there
are several more issues to solve now that it is part of the project
code. In my README docs I will place the initial versions of this code
for reference.