Change how the cart is stored into a list of objects
This was done to allow the storing of the customized invite data
within the shopping cart. Now we will be able to let the customer
edit their changes from within the shopping cart and also pass
this data on to the checkout process.

All cart views which altered the cart have been updated to use the
object list, as well as updating the contexts.py file.