Fix error with the copy of the invite fields array
The problem was caused by using a simple '=' rather than make a
copy, this creates a reference.  This problem was causing the full
reset button on the custom invite not to work correctly. Found help at
(https://www.samanthaming.com/tidbits/35-es6-way-to-clone-an-array/)
to help solve the problem using a shallow copy.